@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 22
-    buildToolsVersion "21.1.2"
+    compileSdk 34
+    buildToolsVersion = "34.0.0"
 
     defaultConfig {
         applicationId "gepalcreations.canwemeet"
-        minSdkVersion 17
-        targetSdkVersion 22
-        versionCode 6
-        versionName "0.5"
+        minSdkVersion 27
+        targetSdkVersion 35
+        versionCode 7
+        versionName "0.6"
     }
     buildTypes {
         release {
@@ -20,7 +20,6 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:22.1.1'
-    compile 'joda-time:joda-time:2.7' // for joda
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.android.support:appcompat-v7:28.0.0'
 }
